+++ Sources/KeyboardKit/Previews/KeyboardPreviews.swift
@@ -1,0 +1,12 @@
+//
+//  KeyboardPreviews.swift
+//  KeyboadKit
+//
+//  Created by Daniel Saidi on 2023-08-08.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/// This is a namespace for various preview-related types.
+public struct KeyboardPreviews {}