+++ Sources/KeyboardKit/_Deprecated/Callouts+Deprecated.swift
@@ -44,3 +44,6 @@ public typealias ActionCalloutStyle = KeyboardActionCalloutStyle
 
 @available(*, deprecated, renamed: "KeyboardInputCalloutStyle")
 public typealias InputCalloutStyle = KeyboardInputCalloutStyle
+
+@available(*, deprecated, message: "This will be removed in KeyboardKit 8.0. Use CalloutActionProvider & LocalizedService instead")
+public typealias LocalizedCalloutActionProvider = CalloutActionProvider & LocalizedService