+++ Sources/KeyboardKit/_Deprecated/DisabledAudioFeedbackEngine.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-@available(*, deprecated, message: "This engine is no longer used.")
-public class DisabledAudioFeedbackEngine: AudioFeedbackEngine {
-
-    public override func trigger(_ audio: AudioFeedback) {}
-}