+++ Sources/KeyboardKit/_Deprecated/InputSetProviderBased.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-@available(*, deprecated, message: "Use input sets directly instead.")
-public protocol InputSetProviderBased {
-
-    func register(inputSetProvider: InputSetProvider)
-}