+++ Sources/KeyboardKit/_Deprecated/KeyboardAction+Deprecated.swift
@@ -14,6 +14,12 @@ public extension KeyboardAction {
     }
 }
 
+@available(*, deprecated, message: "This protocol is deprecated.")
+public protocol KeyboardActionMappable {
+
+    var keyboardAction: KeyboardAction { get }
+}
+
 @available(*, deprecated, renamed: "KeyboardAction.Rows")
 public typealias KeyboardActionRows = KeyboardAction.Rows
 
@@ -54,3 +60,18 @@ private extension String {
         return description
     }
 }
+
+#if os(iOS) || os(tvOS)
+import UIKit
+
+@available(*, deprecated, message: "This extension is deprecated.")
+extension UIReturnKeyType: KeyboardActionMappable {}
+
+public extension UIReturnKeyType {
+
+    @available(*, deprecated, message: "This extension is deprecated.")
+    var keyboardAction: KeyboardAction {
+        .primary(keyboardReturnKeyType)
+    }
+}
+#endif