+++ Sources/KeyboardKit/_Deprecated/LocalizedInputSetProvider.swift
@@ -1,4 +1,0 @@
-import Foundation
-
-@available(*, deprecated, message: "Use input sets directly instead.")
-public typealias LocalizedInputSetProvider = InputSetProvider & LocalizedService