+++ Sources/KeyboardKit/_Deprecated/Previews+Deprecated.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+@available(*, deprecated, renamed: "KeyboardPreviews.ActionHandler")
+public typealias PreviewKeyboardActionHandler = KeyboardPreviews.ActionHandler
+
+@available(*, deprecated, renamed: "KeyboardPreviews.InputViewController")
+public typealias PreviewKeyboardInputViewController = KeyboardPreviews.InputViewController
+
+@available(*, deprecated, renamed: "KeyboardPreviews.LayoutProvider")
+public typealias PreviewKeyboardLayoutProvider = KeyboardPreviews.LayoutProvider
+
+@available(*, deprecated, renamed: "KeyboardPreviews.StyleProvider")
+public typealias PreviewKeyboardStyleProvider = KeyboardPreviews.StyleProvider
+
+@available(*, deprecated, renamed: "KeyboardPreviews.TextDocumentProxy")
+public typealias PreviewTextDocumentProxy = KeyboardPreviews.TextDocumentProxy