+++ Sources/KeyboardKit/_Deprecated/StandardAudioFeedbackEngine.swift
@@ -1,6 +1,0 @@
-#if os(iOS) || os(macOS) || os(tvOS)
-import AudioToolbox
-
-@available(*, deprecated, renamed: "AudioFeedbackEngine")
-open class StandardAudioFeedbackEngine: AudioFeedbackEngine {}
-#endif