+++ Sources/KeyboardKit/_Deprecated/StandardKeyboardFeedbackHandler+Deprecated.swift
@@ -1,9 +1,0 @@
-import Foundation
-
-@available(*, deprecated, message: "Use settings directly instead")
-public extension StandardKeyboardFeedbackHandler {
-    
-    var audioConfig: AudioFeedbackConfiguration { settings.audioConfiguration }
-    
-    var hapticConfig: HapticFeedbackConfiguration { settings.hapticConfiguration }
-}